tests/requirements.txt
mypy
pre-commit
pytest-cov
pytest-html
ruff
typing-extensions
mypy_extensions
types-mypy-extensions
lxml
click
types-click